Fill in test coverage for missing branches.
