+++ src/test/java/net/jodah/failsafe/PredicatesTest.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+package net.jodah.failsafe;
+
+import net.jodah.failsafe.function.BiPredicate;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import static org.testng.Assert.assertEquals;
+
+@Test
+public class PredicatesTest {
+  public void testResultPredicateOnlyHandlesResults() {
+    BiPredicate<Object, Throwable> resultPredicate = Predicates.resultPredicateFor(result -> true);
+    Assert.assertTrue(resultPredicate.test("result", null));
+    Assert.assertFalse(resultPredicate.test(null, new RuntimeException()));
+  }
+}
